requirements.txt
pandas
requests
selenium
psycopg2-binary